Removed an experimental ByteArray implementation.
The idea behind the experimental array was to provide a simple, versatile ByteArray implementation, which would be an independent on STL. The problem with STL is that there are minor differences between the STL implementations available on the systems which still may cause a quite problematic incompatibilities. Unfortunatelly, I don't have enough spare time to finish this custom ByteArray and therefore I'm switching to pure STL implementation now. The vector based ByteArray is much more comfortable for a regular use, so it won the competition.